Extract session form into it's own component
